StationsControl/StationControl/pyproject.toml
[tool.poetry]
name = "stationcontrol"
version = "0.1.0"
description = ""
authors = ["Tanusha <tanusha20020@tut.by>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
django-environ = "^0.9.0"
django-multiselectfield = "^0.1.12"
pillow = "^9.3.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.24.2"}
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
pytest = "^7.2.0"
coverage = "^6.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
